backend/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.6
